requirements.txt
flask
requests
google-generativeai
gunicorn
python-dotenv